fix: handle assembly name param in translator, update tests
