+++ pyproject.toml
@@ -20,7 +20,7 @@ dependencies = [
     "fastapi>=0.95.0",
     "python-multipart",  # required for fastapi file uploads
     "uvicorn",
-    "ga4gh.vrs[extras]>=2.3.1,<3.0",
+    "ga4gh.vrs[extras]>=2.3.2,<3.0",
     "sqlalchemy~=2.0.0",
     "snowflake-sqlalchemy~=1.8.2",
     "pyyaml",
